pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = ["Vahid Al <thevahidal@gmail.com>"]
readme = "README.md"
packages = [{include = "url_shortener"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
